ai-code-debugger/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
google-generativeai==0.3.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
docker==7.0.0
groq
